[PC-555] GetOpponentProfileUseCase 로직을 Repository 수준으로 올림
