+++ core/data/src/main/java/com/puzzle/data/repository/MatchingRepositoryImpl.kt
@@ -1,6 +1,8 @@
 package com.puzzle.data.repository
 
+import com.puzzle.common.suspendRunCatching
 import com.puzzle.domain.model.match.MatchInfo
+import com.puzzle.domain.model.profile.OpponentProfile
 import com.puzzle.domain.model.profile.OpponentProfileBasic
 import com.puzzle.domain.model.profile.OpponentValuePick
 import com.puzzle.domain.model.profile.OpponentValueTalk
@@ -11,6 +13,8 @@ import com.puzzle.network.model.matching.GetOpponentProfileImageResponse
 import com.puzzle.network.model.matching.GetOpponentValuePicksResponse
 import com.puzzle.network.model.matching.GetOpponentValueTalksResponse
 import com.puzzle.network.source.matching.MatchingDataSource
+import kotlinx.coroutines.async
+import kotlinx.coroutines.coroutineScope
 import javax.inject.Inject
 
 class MatchingRepositoryImpl @Inject constructor(
@@ -26,19 +30,48 @@ class MatchingRepositoryImpl @Inject constructor(
     override suspend fun getMatchInfo(): Result<MatchInfo> = matchingDataSource.getMatchInfo()
         .mapCatching(GetMatchInfoResponse::toDomain)
 
-    override suspend fun getOpponentValueTalks(): Result<List<OpponentValueTalk>> =
+    override suspend fun getOpponentProfile(): Result<OpponentProfile> = suspendRunCatching {
+        coroutineScope {
+            val valueTalksDeferred = async { getOpponentValueTalks() }
+            val valuePicksDeferred = async { getOpponentValuePicks() }
+            val profileBasicDeferred = async { getOpponentProfileBasic() }
+            val profileImageDeferred = async { getOpponentProfileImage() }
+
+            val valuePicks = valuePicksDeferred.await().getOrThrow()
+            val valueTalks = valueTalksDeferred.await().getOrThrow()
+            val profileBasic = profileBasicDeferred.await().getOrThrow()
+            val imageUrl = profileImageDeferred.await().getOrThrow()
+
+            OpponentProfile(
+                description = profileBasic.description,
+                nickname = profileBasic.nickname,
+                age = profileBasic.age,
+                birthYear = profileBasic.birthYear,
+                height = profileBasic.height,
+                weight = profileBasic.weight,
+                location = profileBasic.location,
+                job = profileBasic.job,
+                smokingStatus = profileBasic.smokingStatus,
+                valuePicks = valuePicks,
+                valueTalks = valueTalks,
+                imageUrl = imageUrl,
+            )
+        }
+    }
+
+    private suspend fun getOpponentValueTalks(): Result<List<OpponentValueTalk>> =
         matchingDataSource.getOpponentValueTalks()
             .mapCatching(GetOpponentValueTalksResponse::toDomain)
 
-    override suspend fun getOpponentValuePicks(): Result<List<OpponentValuePick>> =
+    private suspend fun getOpponentValuePicks(): Result<List<OpponentValuePick>> =
         matchingDataSource.getOpponentValuePicks()
             .mapCatching(GetOpponentValuePicksResponse::toDomain)
 
-    override suspend fun getOpponentProfileBasic(): Result<OpponentProfileBasic> =
+    private suspend fun getOpponentProfileBasic(): Result<OpponentProfileBasic> =
         matchingDataSource.getOpponentProfileBasic()
             .mapCatching(GetOpponentProfileBasicResponse::toDomain)
 
-    override suspend fun getOpponentProfileImage(): Result<String> =
+    private suspend fun getOpponentProfileImage(): Result<String> =
         matchingDataSource.getOpponentProfileImage()
             .mapCatching(GetOpponentProfileImageResponse::toDomain)
