+++ core/domain/src/main/java/com/puzzle/domain/repository/MatchingRepository.kt
@@ -1,19 +1,14 @@
 package com.puzzle.domain.repository
 
 import com.puzzle.domain.model.match.MatchInfo
-import com.puzzle.domain.model.profile.OpponentProfileBasic
-import com.puzzle.domain.model.profile.OpponentValuePick
-import com.puzzle.domain.model.profile.OpponentValueTalk
+import com.puzzle.domain.model.profile.OpponentProfile
 
 interface MatchingRepository {
     suspend fun reportUser(userId: Int, reason: String): Result<Unit>
     suspend fun blockUser(userId: Int): Result<Unit>
     suspend fun blockContacts(phoneNumbers: List<String>): Result<Unit>
     suspend fun getMatchInfo(): Result<MatchInfo>
-    suspend fun getOpponentValueTalks(): Result<List<OpponentValueTalk>>
-    suspend fun getOpponentValuePicks(): Result<List<OpponentValuePick>>
-    suspend fun getOpponentProfileBasic(): Result<OpponentProfileBasic>
-    suspend fun getOpponentProfileImage(): Result<String>
+    suspend fun getOpponentProfile(): Result<OpponentProfile>
     suspend fun checkMatchingPiece(): Result<Unit>
     suspend fun acceptMatching(): Result<Unit>
 }
